Split verifier derivation out of oversized runtime file
CI correctly rejected autonomous_coding_jobs_runtime.rs after the repo-aware verifier work pushed the file over the 4000-line production threshold. The verifier derivation helpers now live in a focused runtime submodule and the new 3810 tests live under the test-only path that the oversized-file guard intentionally skips.

Constraint: Quality CI enforces .github/scripts/oversized_file_guard.py with a 4000-line default threshold
Rejected: Add an oversized-file exemption | this new code can be split cleanly and should not expand an existing hotspot
Confidence: high
Scope-risk: narrow
Directive: Keep future autonomous coding job additions in focused submodules instead of growing autonomous_coding_jobs_runtime.rs
Tested: cargo fmt --check
Tested: CARGO_TARGET_DIR=/tmp/rust_pi-3810-target cargo test -p tau-runtime spec_3810 -- --test-threads=1
Tested: CARGO_TARGET_DIR=/tmp/rust_pi-3810-target cargo test -p tau-runtime autonomous_coding_jobs_runtime -- --test-threads=1
Tested: CARGO_TARGET_DIR=/tmp/rust_pi-3810-target cargo clippy -p tau-runtime --lib --tests -- -D warnings
Tested: python3 .github/scripts/oversized_file_guard.py --repo-root . --default-threshold 4000 --exemptions-file tasks/policies/oversized-file-exemptions.json --policy-guide docs/guides/oversized-file-policy.md --json-output-file /tmp/rust_pi-3810-oversized-file-guard.json
Tested: git diff --check
Not-tested: Full GitHub CI after this commit before push
